Add coloured terminal support
Introduce logError, logWarning and logInfo, correspondingly.

Adjust the spinner to automatically log error on exceptoins.

Introduce logErrorAndExit that uses exception (thus cleans up the
spinner) in combination with `withCLI` that ensures that the error
message is printed even if it was invoked from outside of any
spinner.

Wherever possible use logError/logWarning. Some uses of `fail`
are left as-is for now, but they will need addressing eventually.