Testing static render only for head
Currently there is no fancy switchover from the static html to the
dynamic rendered frontend so obelisk-frontend clears head and body and
rerenders them through js. When server rendering the script tags in
head would block before the contents of the body gets rendered which
is required for actions in body that depend on the libraries being
loaded, but when rendering through js it breaks the assurance that
script elements would block till done loading introducing wrong/odd
behavior for some js libs that must be loaded before rendering the
next element.

In most cases head must be static for which the switchover behavior is
not desired and even leads to wrong behavior.