Fix dep/.gitignore rule to allow new dep’s
We should allow empty directories here so that the directory shows up
as "Untracked files" in “git status“. Otherwise, it is possible to
forget about some newly added untracked dependencies.

Git apparently will not show ignored directories even if the files in
the directory are specifically unignored. This kind of looks like a
bug.