Allow virtualization types other than Amazon.
When deploying to non amazon instances an error related to GRUB was appearing.

This commit creates a new option virtualization_type that defaults to `amazon` but can be replaced by `virtualbox`, `azure`, `brightbox`, `digital-ocean`, `docker-image` and `google-compute` image.

This behavior is not documented because no options other than amazon are formally supportedi in Obelisk, but makes an implicit assumption explicit.

This has been reported on Github and Reddit:

https://www.reddit.com/r/reflexfrp/comments/9yzqok/obelisk_deployment_issues/
and
https://github.com/obsidiansystems/obelisk/issues/444