Make obelisk depend on reflex-dom-core instead of reflex-dom
All of the Obelisk libraries only need reflex-dom-core to work.
reflex-dom reexports reflex-dom-core, so this worked fine before, but
it had the side effect of pulling in jsaddle-webkit2gtk. For apps that
are just using jsaddle-warp, this pulls in a lot of GTK dependencies
that are never used. Switching everything to reflex-dom-core allows
individual apps avoid these extra dependencies.