Add missing changelog entry for obelisk 0.3 routes change
In b6da186353c525aaf674e16933b48311dca5a1d3, we added a new FullRoute
structure that contains the frontend, backend, and missing routes.
This included changes to the skeleton, but was missing a note in the
changelog that previous apps based off of the skeleton (or really any
obelisk apps) needed to be updated.

This change adds the missing note to the changelog. Emphasis is made
that this is a breaking change that requires some changes on the
user’s part. Some idea of what the user needs to do is given as well.
I have also included an example for what I had to do in
reflex-examples to get it working with Obelisk routes.