Add ob hoogle documentation to obelisk readme
Remove editorialising the user experience of routes and its comparison to other routing
packages.

Remove the use of the newtype and stick to a bare Int for now.

Remove and instead link to the obelisk documentation for the live development environment.

Remove the aside for the wrapped typeclass for a later section.