Fixed up some wording, created tentative class for ReadShow encoders
Created the CanShowReadSafe class intended for types that have Read and Show instances
that are inverses of each other.

Added an exported testing function that is used for the existing instances and users can
use to try to validate a type of their own.

Should it use Text instead of String?