Remove leading slash in generated asset paths
[#831](https://github.com/obsidiansystems/obelisk/pull/831) changed the
way that we generate routes for static assets. The old approach would
place the assets at a path starting with "static". The new approach
added a leading slash, which, when rendered in URLs, would lead to
double slashes at the beginning of routes.

This change brings the new behavior in line with the old behavior.