Invoke shells in a more pure way; escape ghcid commands differently
System.Process invokes `shell` commands using "/bin/sh". Uh-oh! That's
not what we want at all. So, instead, we manually invoke our shells
explicitly using `bash` that is in the environment from the nix-shell.

However, we cannot control the executables we depend on, like ghcid,
which also use System.Process. Since that is the case, we need to fall
back to more generic, but less "surefire" escaping, that works for all
bourne-style shells.