Implement route links which respond to modifier keys
This implementation falls back to the default behavior of the browser
when a modifier key is active while a route link is clicked. It is not
possible to know what behavior is expected by the user for their
particular browser, but there is certainly an expectation.

This implementation does not completely suffice as a replacement for the
current route link behavior, because it does not compose well across
Hydration. Hydration will need to be extended before this implementation
can be prerendered properly, in a manner composable with the
user-supplied child widget that is rendered within the route link's
anchor tag.