Remove NIXPATH references; bump gitignore.nix
Since we upgraded obelisk to use nix-thunk, which was also recently
upgraded to avoid referencing the NIXPATH via `<nixpkgs>`, we can make
obelisk just a little purer by upgrading the thunks. In the tests, we
use a pinned nixpkgs so that even our VM tests do not have this
impurity. The self-tests for legacy thunks, which necessarily use
`<nixpkgs>` are now the only references to it in the code.

gitignore.nix upstream has also recently fixed a source of impurity, so
we bump that dependency.

This is a modest step towards pure-eval and eventually, flakes support.