Batch build all hash file derivaitons before importing outputs
This avoids terrible roundtrip latency we have before, where we kick off
and wait for each build one by one (due to the nature of the evalutor).

If we had a very smart concurrent evaluator we wouldn't need to do this,
it could "discover" these are in paralell on its own. But we don't so we
better do this in the meantime.