command: Use relative paths in getGitHash
This seems to be the only part of the code that doesn't expect a relative path to the current working directory, and that's a quirk of `git rev-parse` with a revision specified. (Without `HEAD`, this would also work, but it would work with files not in the repository.)

This change ensures we are always looking at files in the current directory so deployment files do not have to be located in the root of a repository.